# Direct_Debit/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: Swagger for Real-time Direct Debit API Specification
  title: Real-time Direct Debit API Specification for HK CMB
  version: 1.0.6
servers:
- description: eDDA mock endpoints
  url: https://hsbcdeveloperportalpreprod.digitalapicraft.com:8085/https://sandbox.ob.business.hsbc.com.hk/mock/open-banking/v1.0/direct-debit
  variables:
    version:
      default: v1.0
- description: eDDA endpoints
  url: https://secure.sandbox.ob.business.hsbc.com.hk/open-banking/v1.0/direct-debit
  variables:
    version:
      default: v1.0
paths:
...
  /reports:
    post:
      operationId: postreports
      parameters:
      - description: "Standard HTTP header that allows credentials to be provided\
          \ to the authorisation/resource server. Based on the OAuth 2.0/OIDC framework,\
          \ this consists of basic or bearer authentication schemes."
        example: Bearer LbGDbcFVVfE-ngpjj8KaH5wrEfg
        explode: false
        in: header
        name: Authorization
        required: true
        schema:
          type: string
        style: simple
      - description: The PSU's IP address if the PSU is currently logged in with the
          TPP.
        explode: false
        in: header
        name: x-fapi-customer-ip-address
        required: false
        schema:
          type: string
        style: simple
      - description: "The time when the PSU last logged in with the TPP. All dates\
          \ in the HTTP headers are represented as RFC 7231 Full Dates. An example\
          \ is below: Sun, 10 Sep 2017 19:43:31 UTC pattern: '^(Mon|Tue|Wed|Thu|Fri|Sat|Sun),\
          \ \\d{2} (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec) \\d{4} \\d{2}:\\\
          d{2}:\\d{2} (GMT|UTC)$'"
        explode: false
        in: header
        name: x-fapi-auth-date
        required: false
        schema:
          type: string
        style: simple
      - description: An RFC4122 UID used as a correlation id.
        explode: false
        in: header
        name: x-fapi-interaction-id
        required: false
        schema:
          type: string
        style: simple
      - description: |
          Every request will be processed only once per x-idempotency-key.  The
          Idempotency Key will be valid for 24 hours.
        example: LbGpjj8KaH5wrEfg
        explode: false
        in: header
        name: x-idempotency-key
        required: true
        schema:
          maxLength: 40
          pattern: ^(?!\s)(.*)(\S)$
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reportingRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/reportingResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorInfo'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorInfo'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorInfo'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorInfo'
          description: Not Found
        "405":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorInfo'
          description: Method Not Allowed
        "406":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorInfo'
          description: Not Acceptable
        "415":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorInfo'
          description: Unsupported Media Type
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorInfo'
          description: Internal Server Error
      security:
      - bearerAuthSession: []
      tags:
      - Fulfilment endpoints
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Fulfilment endpoints
components:
...
  schemas:
...
    errorInfo:
      additionalProperties: false
      description: "An array of detail error codes, and messages, and URLs to documentation\
        \ to help remediation."
      example:
        id: id
        errors:
        - extendedDetails:
            path: path
          code: code
          causes: causes
        - extendedDetails:
            path: path
          code: code
          causes: causes
      properties:
        id:
          description: "A unique reference for the error instance, for audit purposes,\
            \ in case of unknown/unclassified errors."
          maxLength: 40
          minLength: 1
          type: string
        errors:
          items:
            $ref: '#/components/schemas/HKError1'
          minItems: 1
          type: array
      required:
      - code
      - errors
      - message
      type: object
...
    HKError1:
      additionalProperties: false
      example:
        extendedDetails:
          path: path
        code: code
        causes: causes
      minProperties: 1
      properties:
        code:
          description: "Low level textual error code, e.g., HK.HKMA.Field.Missing"
          type: string
          x-namespaced-enum:
          - OB.Field.Expected
          - OB.Field.Invalid
          - OB.Field.InvalidDate
          - OB.Field.Missing
          - OB.Field.Unexpected
          - OB.Field.InvalidHeader
          - OB.Field.InvalidDateFormat
          - OB.Field.InvalidDateRange
          - OB.UnexpectedError
          - OB.InternalError
          - OB.InvalidDateRange
          - OB.InvalidDateFormat
          - OB.InvalidRequestParams
          - OB.UnexpectedParamsNo
          - OB.ParamsDuplicated
          - OB.MissingAccountNumber
          - OB.BadRequest
        causes:
          description: |-
            A description of the error that occurred. e.g., 'A mandatory field isn't supplied' or 'RequestedExecutionDateTime must be in future'
            HKMA doesn't standardise this field
          maxLength: 500
          minLength: 1
          type: string
        extendedDetails:
          $ref: '#/components/schemas/HKError1_extendedDetails'
      required:
      - causes
      - code
      type: object
...
    processResult:
      example:
        rejectReasonList:
        - rejectReasonInfomation: rejectReasonInfomation
          rejectReasonCode: rejectReasonCode
        - rejectReasonInfomation: rejectReasonInfomation
          rejectReasonCode: rejectReasonCode
        responseCode: "00"
      properties:
        responseCode:
          description: 'Response code '
          enum:
          - "00"
          - "99"
          maxLength: 4
          type: string
        rejectReasonList:
          items:
            $ref: '#/components/schemas/rejectReasonEntry'
          type: array
      required:
      - responseCode
      type: object
...
    cashAccount:
      properties:
        bankCode:
          description: Bank Code.
          maxLength: 3
          pattern: "^[0-9]{3,3}"
          type: string
        accountIdentification:
          description: Account number / Schema ID.No hyphen full stop nor special
            character.
          maxLength: 35
          minLength: 1
          pattern: "^[a-zA-Z0-9]{1,35}"
          type: string
        currency:
          description: Account currency.
          maxLength: 3
          pattern: "^[A-Z]{3,3}"
          type: string
        accountSchemeName:
          description: Scheme code.If not provided \BBAN\ will be the default value.
          enum:
          - BBAN
          maxLength: 4
          type: string
        accountSubType:
          description: "Specifies the sub type of account (product family group).\
            \ Channel specific description. This is the field where the response for\
            \ cashAccountType field will be mapped as per STET standard . For greece\
            \ and Malta, this field will be used to map the product category code"
          example: "Current Account, Savings Account, Student Account, CACC, CARD,\
            \ DDA – Current & Savings Accounts or CDA – Term Deposits or ILA – Loans\
            \ or CCA – Credit Cards or OTH - Other"
          maxLength: 35
          type: string
      required:
      - accountIdentification
      - bankCode
      - currency
      type: object
    rejectReasonEntry:
      description: Reject Reason List-Entry
      example:
        rejectReasonInfomation: rejectReasonInfomation
        rejectReasonCode: rejectReasonCode
      properties:
        rejectReasonCode:
          description: "Reject Reason List-Entry-Element: Code."
          maxLength: 10
          type: string
        rejectReasonInfomation:
          description: "Reject Reason List-Entry-Element: Information"
          maxLength: 105
          type: string
      type: object
...
    reportingResponse:
      example:
        reportContent: reportContent
        processResult:
          rejectReasonList:
          - rejectReasonInfomation: rejectReasonInfomation
            rejectReasonCode: rejectReasonCode
          - rejectReasonInfomation: rejectReasonInfomation
            rejectReasonCode: rejectReasonCode
          responseCode: "00"
      properties:
        processResult:
          $ref: '#/components/schemas/processResult'
        reportContent:
          description: Report Content.Encoded in BASE64.
          format: byte
          type: string
      required:
      - processResult
      - reportContent
      type: object
    reportingRequest:
      properties:
        reportType:
          description: Type of report.
          enum:
          - DDAReport
          - DDIReport
          maxLength: 10
          type: string
        reportingDate:
          description: Date of report.
          format: date
          type: string
        creditorAccount:
          $ref: '#/components/schemas/cashAccount'
      required:
      - creditorAccount
      - reportType
      - reportingDate
      type: object
...
    HKError1_extendedDetails:
      description: Provide details for the error
      example:
        path: path
      properties:
        path:
          description: "Recommended but optional reference to the JSON Path of the\
            \ field with error, e.g., Data.Initiation.InstructedAmount.Currency"
          maxLength: 500
          minLength: 1
          type: string
      type: object
  securitySchemes:
    apiOAuth:
      flows:
        implicit:
          authorizationUrl: https://api.hsbc.com/.../...
          scopes:
            "-''": ""
      type: oauth2
    bearerAuthSession:
      bearerFormat: JWT
      description: |
        **Sample Value** ``` eyJraWQiOiI1M2Q4NjFjNC1iZTMwLTQzNDItYmU0ZS05N2RlYzQ5YmZhZTIiLCJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJqdGkiOiI3YjA1NWY4YS01YmJmLTQxZjgtOTYwMC05MTFjNTgyMzJiNjUiLCJpYXQiOjE1MTU3NjE2NzEsInJlc3BvbnNlVHlwZSI6ImNvZGUiLCJpbnRlbnRUeXBlIjoiYWNjb3VudCIsImNsaWVudElkIjoiNTZkZmM2ZTMtMzQ3Yi00ZWY1LTgzY2EtMDcxMDQxZWIwMDMxIiwicmVkaXJlY3RVcmwiOiJodHRwczovLzEyNy4wLjAuMS9yZWRpcmVjdDEiLCJjdXN0b21lcklwIjoiMTI3LjAuMC4xIiwiYnJhbmRJZCI6ImhzYmMiLCJpbnRlbnRJZCI6ImJhM2I0NzY2LTQyNzktNDhjZC05ZGYzLWIyYmQ1NGJhZmNkNiIsImNsYWltSWQiOiI4YjY5ZmM2Yjc2NGU3YTllNTk2YmI2OWMzMzI5NDYxNzIyM2JhMDE3NTAxNTIyN2E5N2M5YzEyZWUzM2JhNWMwIiwic3RhdGUiOiJzb21lLXN0YXRlLTEyMzQ1Njc4Iiwibm9uY2UiOiJuLTBTNl9XekEyTWoifQ.1nqXILi5L1HcMmppIavgd8S6VVhDNBeF-i0XrGwFxVaAs3NW8BuReMRFEbsPC0p56DnXQ3gvYA-BbJ-ahS0vRiMvrxa3VEtTOswJBtjpLBphfOkUvgIAthB2MFv7dVsZvBs9Kqfqis5MINPjafSXaFh9ZelZyu1G_zaQb6LvcxMce5b1fBtFFV3gLGCE9ScdeSMC08mZqPajeW7mP37wBSp_QMSmxeK89KrmZtIMJmte_FuTw-TQ4dwmP6L6nscRAAoTpEaDItxKAWXFyOQPPo2fxZ-79x-Rr12MkTJzoFJbquAwaGEApLgz0JUtp_L20orgrIh8jcY6e7cEt4qglg ``` --- **Decoded**
         ```
          Header
          {
            "kid": "53d861c4-be30-4342-be4e-97dec49bfae2",
            "typ": "JWT",
            "alg": "RS256",
            "jku": ""https://private.api.open.banking.hsbc.com/.well-known/jwks.json"
          }
          Payload
          {
             "jti": "7b055f8a-5bbf-41f8-9600-911c58232b65",
             "iat": "1515761671",
             "scope": "openid aisp or pisp or cbpii or AIS:  32",
             "state": "some-state-12345678",
             "nonce": "n-0S6_WzA2Mj",
             "iss": "https://private.api.personal.hsbc.co.uk",
             "aud": "https://channelUI.com",
             "exp": "1 minute from set time",
             "brand": "PERSONAL",
             "country": "UK",
             "standard": "obie",
             "consentId": "ba3b4766-4279-48cd-9df3-b2bd54bafcd6",
             "consentType": "AIS or PIS or CBPII",
             "customerIp": "91.10.21.5",
             "tppId": "PL-TST-734519779079",
             "tppName" : "My Tpp"
          }
         ``
      scheme: bearer
...